+++ packages/utils/src/jsonPrint.ts
@@ -1,0 +1,11 @@
+export const jsonPrint = (value: any, indent = 2) =>
+  JSON.stringify(
+    value,
+    (_, v) =>
+      typeof v === "bigint"
+        ? `${v}n`
+        : typeof v === "object" && typeof v?.asHex === "function"
+          ? v.asHex()
+          : v,
+    indent,
+  )