+++ packages/utils/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+- Add `jsonPrint` function
+
 ## 0.1.2 - 2024-10-05
 
 ### Fixed